Move MsalMetricsCatalog to Extensibiliry namespace
